feat(go.d.plugin): add dyncfg vnodes (#19205)
* add dyncfg vnodes

* fix tests
