+++ src/go/plugin/go.d/agent/jobmgr/manager.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"log/slog"
 	"os"
+	"strings"
 	"sync"
 	"time"
 
@@ -18,6 +19,7 @@ import (
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/confgroup"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/functions"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/module"
+	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/vnodes"
 
 	"github.com/mattn/go-isatty"
 	"gopkg.in/yaml.v2"
@@ -34,8 +36,9 @@ func New() *Manager {
 		FileLock:        noop{},
 		FileStatus:      noop{},
 		FileStatusStore: noop{},
-		Vnodes:          noop{},
 		FnReg:           noop{},
+
+		Vnodes: make(map[string]*vnodes.VirtualNode),
 
 		discoveredConfigs: newDiscoveredConfigsCache(),
 		seenConfigs:       newSeenConfigCache(),
@@ -64,8 +67,8 @@ type Manager struct {
 	FileLock        FileLocker
 	FileStatus      FileStatus
 	FileStatusStore FileStatusStore
-	Vnodes          Vnodes
 	FnReg           FunctionRegistry
+	Vnodes          map[string]*vnodes.VirtualNode
 
 	discoveredConfigs *discoveredConfigs
 	seenConfigs       *seenConfigs
@@ -90,8 +93,14 @@ func (m *Manager) Run(ctx context.Context, in chan []*confgroup.Group) {
 
 	m.FnReg.Register("config", m.dyncfgConfig)
 
+	//m.dyncfgVnodeModuleCreate()
+	//
+	//for _, cfg := range m.Vnodes {
+	//	m.dyncfgVnodeJobCreate(cfg, dyncfgRunning)
+	//}
+
 	for name := range m.Modules {
-		m.dyncfgModuleCreate(name)
+		m.dyncfgCollectorModuleCreate(name)
 	}
 
 	var wg sync.WaitGroup
@@ -134,7 +143,7 @@ func (m *Manager) run() {
 			case <-m.ctx.Done():
 				return
 			case fn := <-m.dyncfgCh:
-				m.dyncfgConfigExec(fn)
+				m.dyncfgCollectorSeqExec(fn)
 			}
 		} else {
 			select {
@@ -145,7 +154,14 @@ func (m *Manager) run() {
 			case cfg := <-m.rmCh:
 				m.removeConfig(cfg)
 			case fn := <-m.dyncfgCh:
-				m.dyncfgConfigExec(fn)
+				switch id := fn.Args[0]; true {
+				case strings.HasPrefix(id, dyncfgCollectorIDPrefix):
+					m.dyncfgCollectorSeqExec(fn)
+				case strings.HasPrefix(id, dyncfgVnodeIDPrefix):
+					m.dyncfgVnodeSeqExec(fn)
+				default:
+					m.dyncfgRespf(fn, 503, "unknown function '%s' (%s).", fn.Name, id)
+				}
 			}
 		}
 	}
@@ -185,7 +201,7 @@ func (m *Manager) addConfig(cfg confgroup.Config) {
 		ecfg = scfg
 	}
 
-	m.dyncfgJobCreate(ecfg.cfg, ecfg.status)
+	m.dyncfgCollectorJobCreate(ecfg.cfg, ecfg.status)
 
 	if isTerminal || m.PluginName == "nodyncfg" { // FIXME: quick fix of TestAgent_Run (agent_test.go)
 		m.dyncfgConfigEnable(functions.Function{Args: []string{dyncfgJobID(ecfg.cfg), "enable"}})
@@ -274,21 +290,14 @@ func (m *Manager) createCollectorJob(cfg confgroup.Config) (*module.Job, error) 
 		return nil, fmt.Errorf("can not find %s module", cfg.Module())
 	}
 
-	var vnode struct {
-		guid     string
-		hostname string
-		labels   map[string]string
-	}
+	var vnode *vnodes.VirtualNode
 
 	if cfg.Vnode() != "" {
-		n, ok := m.Vnodes.Lookup(cfg.Vnode())
-		if !ok {
+		n, ok := m.Vnodes[cfg.Vnode()]
+		if !ok || n == nil {
 			return nil, fmt.Errorf("vnode '%s' is not found", cfg.Vnode())
 		}
-
-		vnode.guid = n.GUID
-		vnode.hostname = n.Hostname
-		vnode.labels = n.Labels
+		vnode = n
 	}
 
 	m.Debugf("creating %s[%s] job, config: %v", cfg.Module(), cfg.Name(), cfg)
@@ -311,9 +320,9 @@ func (m *Manager) createCollectorJob(cfg confgroup.Config) (*module.Job, error) 
 		IsStock:         cfg.SourceType() == "stock",
 		Module:          mod,
 		Out:             m.Out,
-		VnodeGUID:       vnode.guid,
-		VnodeHostname:   vnode.hostname,
-		VnodeLabels:     vnode.labels,
+	}
+	if vnode != nil {
+		jobCfg.Vnode = *vnode.Copy()
 	}
 
 	job := module.NewJob(jobCfg)
