+++ src/go/plugin/go.d/agent/vnodes/vnodes_test.go
@@ -8,20 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNew(t *testing.T) {
-	assert.NotNil(t, New("testdata"))
-	assert.NotNil(t, New("not_exist"))
-}
-
-func TestVnodes_Lookup(t *testing.T) {
-	req := New("testdata")
-
-	_, ok := req.Lookup("first")
-	assert.True(t, ok)
-
-	_, ok = req.Lookup("second")
-	assert.True(t, ok)
-
-	_, ok = req.Lookup("third")
-	assert.False(t, ok)
+func TestLoad(t *testing.T) {
+	assert.NotNil(t, Load("testdata"))
+	assert.NotNil(t, Load("not_exist"))
 }
