+++ src/go/plugin/go.d/collector/snmp/testdata/config.json
@@ -3,6 +3,7 @@
   "hostname": "ok",
   "create_vnode": true,
   "vnode": {
+    "name": "ok",
     "guid": "ok",
     "hostname": "ok",
     "labels": {
